Address verifier CLI parser review
