+++ ops/scripts/deployment-bus.cjs
@@ -3,6 +3,7 @@
 
 const fs = require("node:fs");
 const path = require("node:path");
+const { parseArgs } = require("./cli-args.cjs");
 
 const SCHEMA_VERSION = "deployment-bus.v1";
 const SHA_RE = /^[0-9a-f]{40}$/i;
@@ -160,40 +161,6 @@ function createPlaywrightPack({
     }),
     artifacts: PLAYWRIGHT_ARTIFACT_PATTERNS,
   });
-}
-
-function parseArgs(argv) {
-  const args = { _: [] };
-  let index = 0;
-
-  while (index < argv.length) {
-    const token = argv[index];
-    if (!token.startsWith("--")) {
-      args._.push(token);
-      index += 1;
-      continue;
-    }
-
-    const rawKey = token.slice(2);
-    const eqIndex = rawKey.indexOf("=");
-    if (eqIndex !== -1) {
-      args[rawKey.slice(0, eqIndex)] = rawKey.slice(eqIndex + 1);
-      index += 1;
-      continue;
-    }
-
-    const nextToken = argv[index + 1];
-    if (!nextToken || nextToken.startsWith("--")) {
-      args[rawKey] = true;
-      index += 1;
-      continue;
-    }
-
-    args[rawKey] = nextToken;
-    index += 2;
-  }
-
-  return args;
 }
 
 function readJson(file) {
